civic_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
PyJWT==2.8.0
supabase==2.3.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6